% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{findCutoff}
\alias{findCutoff}
\title{find cutoff for a 1D distribution}
\usage{
findCutoff(x, returnSil = F, useBL = T, minX = 0)
}
\arguments{
\item{x}{a vector of values.}

\item{returnSil}{Logic, used to specify if the max average silhouette is returned}

\item{useBL}{Logic, used to specify if outliers should be ignored}

\item{minX}{a numerical value, used to specify the min value allowed for the cutoff.}
}
\value{
If returnSil=F, returns a single cutoff value. Otherwise, returns a list containing the cutoff value and the max average silhouette
}
\description{
A function that find cutoff for a 1D distribution.
}

