% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{flowHC}
\alias{flowHC}
\title{cluster cytometry data using hierarchical clustering}
\usage{
flowHC(fcsFrame, excludeClusterParameters, minimumClusterSizePercent = 0.05)
}
\arguments{
\item{fcsFrame}{a flow frame.}

\item{excludeClusterParameters}{A vector specifiying the name of markers not to be used for clustering.}

\item{minimumClusterSizePercent}{a number between 0 and 1, used to specify the minimum size of a cluster relative to all events.}
}
\value{
a list of clusters. Each cluster contains the ID of all cells that belong to the cluster.
}
\description{
A function that cluster cytometry data using hierarchical clustering.
}

