% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{flowSOM.MC}
\alias{flowSOM.MC}
\title{cluster cytometry data using FlowSOM}
\usage{
flowSOM.MC(fcsFrame, excludeClusterParameters)
}
\arguments{
\item{fcsFrame}{a flow frame.}

\item{excludeClusterParameters}{A vector specifiying the name of markers not to be used for clustering.}
}
\value{
a list of clusters. Each cluster contains the ID of all cells that belong to the cluster.
}
\description{
A function that cluster cytometry data using FlowSOM.
}

