% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_functions.R
\name{labelCluster}
\alias{labelCluster}
\title{label each clusters as "+" or "-" or neutral for each marker}
\usage{
labelCluster(fcsFrame, clusterList, excludeClusterParameters = c("TIME"),
  minPercent = 0.05, labelQuantile = 0.95, cutoff = NULL)
}
\arguments{
\item{fcsFrame}{A flowFrame object.}

\item{clusterList}{a list, each element should be a vector containing the IDs of all cells that are belong to a cluster}

\item{excludeClusterParameters}{A vector specifiying the name of markers not to be used for labeling.}

\item{minPercent}{a number between 0 and 0.5. Used to specify the minimum percent of cells in positive and negative region after bisection. Keep it small to avoid bisecting uni-mode distributions.}

\item{labelQuantile}{a number between 0 and 0.5. Used to specify the minimum percent of a cluster required to be larger or smaller than the cutoff value for labeling.}
}
\value{
Returns a list with two components: 1) clusterLabel, contains a vector of lables, each correspond to a cluster in clusterList. 2) cutoff, contains a vector of cutoff values used to bisect each marker.
}
\description{
A function that label each cluster as "+" or "-" or neutral for each marker
}

