% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_functions.R
\name{markerFinder}
\alias{markerFinder}
\title{find markers in a flow frame object}
\usage{
markerFinder(fcsFrame)
}
\arguments{
\item{fcsFrame}{a flow frame object.}
}
\value{
returns a vector of markers.
}
\description{
A function that finds markers in a flow frame object.
}
\details{
If the antibody name is available, the antibody name will be returned, otherwise the channel name will be returned.
}

