% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{metaAnalysis}
\alias{metaAnalysis}
\title{perform meta-analysis}
\usage{
metaAnalysis(value, variableOfInterst, otherVariables, studyID, data, CILevel,
  main, ifScale = c(T, F), cex = 1)
}
\arguments{
\item{value}{a string to specify the column name of the dependent variable (y)}

\item{variableOfInterst}{a string to specify the column name of the independent variable of interest (x1)}

\item{otherVariables}{a string vector to specify the column names of independent variables included in the regression model other than the variableOfInterst.}

\item{studyID}{a string to specify the column name of study ID.}

\item{data}{a data frame containing the data}

\item{CILevel}{a number between 0 to 1, used to specify the confidence interval to be plotted in the forrest plot.}

\item{main}{a string to specify the title of the forrest plot}

\item{ifScale}{a vector of two logic values, specifing if the dependent variable and the variableOfInterst should be scaled when calculating the effect size.}

\item{cex}{a number specifying the amount by which plotting text and symbols should be scaled relative to the default in the forrest plot.}
}
\value{
returns data frame describing the effect size of variableOfInterst on value in each individule studies, as well as the over all effect size.
}
\description{
A function that performs meta-analysis
}
\examples{
#collect all summary statistics
fn=system.file("extdata","",package="MetaCyto")
files=list.files(fn,pattern="cluster_stats_in_each_sample",recursive=T,full.names=T)
fcs_stats=collectData(files,longform=T)
# Collect sample information
files=list.files(fn,pattern="sample_info",recursive=T,full.names=T)
sample_info=collectData(files,longform=F)
# join the cluster summary statistics with sample information
all_data=inner_join(fcs_stats,sample_info,by="fcs_files")

# plot forrest plot to see if the proportion of CCR7+ CD8 T cell
# is affected by age (while controlling for GENDER)
L="CD3+|CD4-|CD8+|CCR7+"
dat=subset(all_data,all_data$parameter_name=="fraction"&
            all_data$label==L)
MA=metaAnalysis(value="value",variableOfInterst="SUBJECT_AGE",main=L,
                otherVariables=c("GENDER"),studyID="study_id",
                data=dat,CILevel=0.95,ifScale=c(T,F))
}

