% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{nameUpdator}
\alias{nameUpdator}
\title{Used to update marker names}
\usage{
nameUpdator(oldNames, newNames, files)
}
\arguments{
\item{oldNames}{a vector of marker names you wish to change}

\item{newNames}{a vector of marker names you wish each oldNames to be changed to.}

\item{files}{a list of "processed_sample_summary.csv" files in which the name change will ocurre.}
}
\value{
Null
}
\description{
A function that update marker names in the files output by the preprocessing.batch function.
}

