% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{panelSummary}
\alias{panelSummary}
\title{summarize markers in panels.}
\usage{
panelSummary(panelInfo, folder, cluster = T, plotImage = T, width = 20,
  height = 20)
}
\arguments{
\item{panelInfo}{a data frame returned by collectData function. It should contain all the information outputted by the preprocessing.batch function.}

\item{folder}{the directory where the output should be written}

\item{cluster}{True or False. Used to indicate if the markers and panels should be clustered in the plot.}

\item{plotImage}{True or False. Used to indicate if a plot summarizing markers in  panels should be produced.}

\item{width}{Used to specify the width of the plot}

\item{height}{Used to specify the height of the plot}
}
\value{
A dataframe describing what markers are in each panel.
}
\description{
A function that summarize markers in panels.
}

