% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{plotGA}
\alias{plotGA}
\title{plot the result from the glmAnalysis function}
\usage{
plotGA(GA, size = 16)
}
\arguments{
\item{GA}{a data frame returned from the function glmAnalysis.}

\item{size}{font size of texts in the plot}
}
\description{
A function that plot the result from the glmAnalysis function.
}

