% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_functions.R
\name{searchCluster.batch}
\alias{searchCluster.batch}
\title{search for clusters using pre-defined labels in cytometry data from different studies in batch}
\usage{
searchCluster.batch(preprocessOutputFolder, outpath = "search_output",
  clusterLabel, ifPlot = T)
}
\arguments{
\item{preprocessOutputFolder}{directory where the pre-processed results are stored.}

\item{outpath}{a string indicating the directory the results should be write to.}

\item{clusterLabel}{a vector containing labels, such as "CD3+|CD4+|CD8-"}

\item{ifPlot}{True or False. Used to specifiy if a the density plot for each cluster should be plotted}
}
\description{
A function that search for clusters using pre-defined labels in cytometry data from different studies in batch.
}
\details{
The function write out the summary statistics for each cluster. A seperate directory will be created for each study.
}

