% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{searchCluster}
\alias{searchCluster}
\title{search for clusters using pre-defined labels}
\usage{
searchCluster(fcsFrame, clusterLabel, cutoff = NULL, rmNull = T,
  preGate = NULL)
}
\arguments{
\item{fcsFrame}{A flowFrame object.}

\item{clusterLabel}{a vector of labels, such as "CD3+|CD4+|CD8-"}

\item{cutoff}{a vector of cutoff values to bisect the distribution of each
marker. The names of the vector should be the same as the marker names. If
NULL, the cutoff value will be determined automatically.}

\item{rmNULL}{True or False. Used to specifiy if a cluster with 0 cells
should be returned or not.}
}
\value{
Returns a list with two components: 1) clusterList, contains a list.
  Each element of the list is a vector containing the ID of all cells in a
  cluster. The names corresponds to the labels specified in clusterLabel. 2)
  cutoff, contains a vector of cutoff values used to bisect each marker.
}
\description{
A function that search for clusters using pre-defined labels
}
\examples{
# Find fcs files
files=system.file("extdata","SDY420/ResultFiles/CyTOF_result",package="MetaCyto")
files=list.files(files,pattern="fcs$",full.names=T)
# Preprocess
fcs = preprocessing(fcsFiles=files,assay ="CyTOF",b=1/8)
# Search clusters
cluster_list=searchCluster(fcsFrame=fcs,
                           clusterLabel=c("CD3+|CD8+","CD3-|CD19+"),
                           cutoff=cluster_label$cutoff)
}

