% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{set2Frame}
\alias{set2Frame}
\title{combine cells in a flow set into a flow frame.}
\usage{
set2Frame(flowSet)
}
\arguments{
\item{flowSet}{a flow set object}
}
\value{
Returns a flowFrame object. All cells from flow set are combined into one flow frame. A new paramter, sample_id, is introduced to indicate the origin of each cell.
}
\description{
A function that combine cells in a flow set into a flow frame.
}

