% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_functions.R
\name{filterLabels}
\alias{filterLabels}
\title{filter cluster labels}
\usage{
filterLabels(labels, minPlus, minMarker, maxMarker)
}
\arguments{
\item{labels}{a vector containing labels for cell clusters}

\item{minPlus}{an integer, used to specify the minmum number of "+" a label should contain.}

\item{minMarker}{an integer, used to specify the minmum number of markers a label should contain.}

\item{maxMarker}{an integer, used to specify the max number of markers a label should contain.}
}
\value{
returns a vector of labels that pass through the filter.
}
\description{
A function that filter cluster labels.
}

