% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{flowSOM.MC}
\alias{flowSOM.MC}
\title{cluster cytometry data using FlowSOM}
\usage{
flowSOM.MC(fcsFrame, excludeClusterParameters, grid = 10, k = 40)
}
\arguments{
\item{fcsFrame}{a flow frame.}

\item{excludeClusterParameters}{A vector specifiying the name of markers not to be used for clustering.}
}
\value{
a list of clusters. Each cluster contains the ID of all cells that belong to the cluster.
}
\description{
A function that cluster cytometry data using FlowSOM.
}
\examples{
# Find fcs files
files=system.file("extdata","SDY420/ResultFiles/CyTOF_result",package="MetaCyto")
files=list.files(files,pattern="fcs$",full.names=T)
# Preprocess
fcs = preprocessing(fcsFiles=files,assay ="CyTOF",b=1/8)
# cluster using flowSOM.MC
cluster_list=flowSOM.MC(fcsFrame=fcs,excludeClusterParameters=c("Time","Cell_length"))
}

