% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaCyto_Functions.R
\name{plotGA}
\alias{plotGA}
\title{plot the result from the glmAnalysis function}
\usage{
plotGA(GA, size = 16)
}
\arguments{
\item{GA}{a data frame returned from the function glmAnalysis.}

\item{size}{font size of texts in the plot}
}
\description{
A function that plot the result from the glmAnalysis function.
}
\examples{
#collect all summary statistics
fn=system.file("extdata","",package="MetaCyto")
files=list.files(fn,pattern="cluster_stats_in_each_sample",recursive=T,full.names=T)
fcs_stats=collectData(files,longform=T)
# Collect sample information
files=list.files(fn,pattern="sample_info",recursive=T,full.names=T)
sample_info=collectData(files,longform=F)
# join the cluster summary statistics with sample information
all_data=inner_join(fcs_stats,sample_info,by="fcs_files")
# See the fraction of what clusters are affected by age (while controlling for GENDER)
GA=glmAnalysis(value="value",variableOfInterst="SUBJECT_AGE",parameter="fraction",
              otherVariables=c("GENDER"),studyID="study_id",label="label",
              data=all_data,CILevel=0.95,ifScale=c(T,F))
GA=GA[order(GA$Effect_size),]
# plot the effect sizes
plotGA(GA)
}

